refactor: move ConnectionDetails to types dir
This is necessary to prevent a circular import when using the
ConnectionDetails class from WebSocketTransport.
